+++ README.md
@@ -65,3 +65,28 @@ and pgpool with:
 
     docker compose up -d pgpool pg18
     PGPORT=7432 go test ./...
+
+You can use PQGO_DEBUG=1 to make the driver print the communication with
+PostgreSQL to stderr; this works anywhere (test or applications) and can be
+useful to debug protocol problems.
+
+For example:
+
+    % PQGO_DEBUG=1 go test -run TestSimpleQuery
+    CLIENT → Startup                 69  "\x00\x03\x00\x00database\x00pqgo\x00user [..]"
+    SERVER ← (R) AuthRequest          4  "\x00\x00\x00\x00"
+    SERVER ← (S) ParamStatus         19  "in_hot_standby\x00off\x00"
+    [..]
+    SERVER ← (Z) ReadyForQuery        1  "I"
+             START conn.query
+             START conn.simpleQuery
+    CLIENT → (Q) Query                9  "select 1\x00"
+    SERVER ← (T) RowDescription      29  "\x00\x01?column?\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x17\x00\x04\xff\xff\xff\xff\x00\x00"
+    SERVER ← (D) DataRow              7  "\x00\x01\x00\x00\x00\x011"
+             END conn.simpleQuery
+             END conn.query
+    SERVER ← (C) CommandComplete      9  "SELECT 1\x00"
+    SERVER ← (Z) ReadyForQuery        1  "I"
+    CLIENT → (X) Terminate            0  ""
+    PASS
+    ok      github.com/lib/pq       0.010s
